xen/vpci: msix: move x86 specific code to x86 file
vpci/msix.c file will be used for arm architecture when vpci msix
support will be added to ARM, but there is x86 specific code in this
file.

Move x86 specific code to the x86/hvm/vmsi.c file to make sure common
code will be used for other architecture.

No functional change intended.

---
(cherry picked from commit 9f36d1b1dffcca1ae3fcb2dfcac4709c39d1b3bc from
 the downstream branch poc/pci-passthrough from
 https://gitlab.com/xen-project/people/bmarquis/xen-arm-poc.git)

changes since cherry-pick:
* rebase on top of the vpci rwlock series
* rebase on top of
  b177892d2d0e ("vpci/msix: handle accesses adjacent to the MSI-X table")

TODO: we are likely not using the (rw)lock appropriately. See TODO in xen/drivers/vpci/msix.c.
