+++ xen/drivers/vpci/msix.c
@@ -17,20 +17,11 @@
  * License along with this program; If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <xen/msi.h>
 #include <xen/sched.h>
 #include <xen/vpci.h>
 
-#include <asm/msi.h>
 #include <asm/p2m.h>
-
-#define VMSIX_ADDR_IN_RANGE(addr, vpci, nr)                               \
-    ((addr) >= vmsix_table_addr(vpci, nr) &&                              \
-     (addr) < vmsix_table_addr(vpci, nr) + vmsix_table_size(vpci, nr))
-
-#define VMSIX_ADDR_SAME_PAGE(addr, vpci, nr)                              \
-    (PFN_DOWN(addr) >= PFN_DOWN(vmsix_table_addr(vpci, nr)) &&            \
-     PFN_DOWN(addr) <= PFN_DOWN(vmsix_table_addr(vpci, nr) +              \
-                                vmsix_table_size(vpci, nr) - 1))
 
 static uint32_t cf_check control_read(
     const struct pci_dev *pdev, unsigned int reg, void *data)
@@ -146,37 +137,6 @@ static void cf_check control_write(
     val = control_read(pdev, reg, data);
     if ( pci_msi_conf_write_intercept(msix->pdev, reg, 2, &val) >= 0 )
         pci_conf_write16(pdev->sbdf, reg, val);
-}
-
-static struct vpci_msix *msix_find(const struct domain *d, unsigned long addr)
-{
-    struct vpci_msix *msix;
-
-    ASSERT(rw_is_locked(&d->pci_lock));
-
-    list_for_each_entry ( msix, &d->arch.hvm.msix_tables, next )
-    {
-        const struct vpci_bar *bars = msix->pdev->vpci->header.bars;
-        unsigned int i;
-
-        for ( i = 0; i < ARRAY_SIZE(msix->tables); i++ )
-            if ( bars[msix->tables[i] & PCI_MSIX_BIRMASK].enabled &&
-                 VMSIX_ADDR_SAME_PAGE(addr, msix->pdev->vpci, i) )
-                return msix;
-    }
-
-    return NULL;
-}
-
-static int cf_check msix_accept(struct vcpu *v, unsigned long addr)
-{
-    int rc;
-
-    read_lock(&v->domain->pci_lock);
-    rc = !!msix_find(v->domain, addr);
-    read_unlock(&v->domain->pci_lock);
-
-    return rc;
 }
 
 static bool access_allowed(const struct pci_dev *pdev, unsigned long addr,
@@ -368,11 +328,10 @@ static int adjacent_read(const struct domain *d, const struct vpci_msix *msix,
     return X86EMUL_OKAY;
 }
 
-static int cf_check msix_read(
-    struct vcpu *v, unsigned long addr, unsigned int len, unsigned long *data)
+int cf_check vpci_msix_read(struct vpci_msix *msix, unsigned long addr,
+    unsigned int len, unsigned long *data)
 {
-    struct domain *d = v->domain;
-    struct vpci_msix *msix;
+    struct domain *d = msix->pdev->domain;
     const struct vpci_msix_entry *entry;
     unsigned int offset;
 
@@ -380,7 +339,10 @@ static int cf_check msix_read(
 
     read_lock(&d->pci_lock);
 
+#if 0
+    /* TODO: caller must lock appropriately */
     msix = msix_find(d, addr);
+#endif
     if ( !msix )
     {
         read_unlock(&d->pci_lock);
@@ -516,17 +478,19 @@ static int adjacent_write(const struct domain *d, const struct vpci_msix *msix,
     return X86EMUL_OKAY;
 }
 
-static int cf_check msix_write(
-    struct vcpu *v, unsigned long addr, unsigned int len, unsigned long data)
+int cf_check vpci_msix_write(struct vpci_msix *msix, unsigned long addr,
+    unsigned int len, unsigned long data)
 {
-    struct domain *d = v->domain;
-    struct vpci_msix *msix;
+    struct domain *d = msix->pdev->domain;
     struct vpci_msix_entry *entry;
     unsigned int offset;
 
     read_lock(&d->pci_lock);
 
+#if 0
+    /* TODO: caller must lock appropriately */
     msix = msix_find(d, addr);
+#endif
     if ( !msix )
     {
         read_unlock(&d->pci_lock);
@@ -626,79 +590,6 @@ static int cf_check msix_write(
     return X86EMUL_OKAY;
 }
 
-static const struct hvm_mmio_ops vpci_msix_table_ops = {
-    .check = msix_accept,
-    .read = msix_read,
-    .write = msix_write,
-};
-
-int vpci_make_msix_hole(const struct pci_dev *pdev)
-{
-    struct domain *d = pdev->domain;
-    unsigned int i;
-
-    if ( !pdev->vpci->msix )
-        return 0;
-
-    /* Make sure there's a hole for the MSIX table/PBA in the p2m. */
-    for ( i = 0; i < ARRAY_SIZE(pdev->vpci->msix->tables); i++ )
-    {
-        unsigned long start = PFN_DOWN(vmsix_table_addr(pdev->vpci, i));
-        unsigned long end = PFN_DOWN(vmsix_table_addr(pdev->vpci, i) +
-                                     vmsix_table_size(pdev->vpci, i) - 1);
-
-        for ( ; start <= end; start++ )
-        {
-            p2m_type_t t;
-            mfn_t mfn = get_gfn_query(d, start, &t);
-
-            switch ( t )
-            {
-            case p2m_mmio_dm:
-            case p2m_invalid:
-                break;
-            case p2m_mmio_direct:
-                if ( mfn_x(mfn) == start )
-                {
-                    p2m_remove_identity_entry(d, start);
-                    break;
-                }
-                /* fallthrough. */
-            default:
-                put_gfn(d, start);
-                gprintk(XENLOG_WARNING,
-                        "%pp: existing mapping (mfn: %" PRI_mfn
-                        "type: %d) at %#lx clobbers MSIX MMIO area\n",
-                        &pdev->sbdf, mfn_x(mfn), t, start);
-                return -EEXIST;
-            }
-            put_gfn(d, start);
-        }
-    }
-
-    if ( is_hardware_domain(d) )
-    {
-        /*
-         * For dom0 only: remove any hypervisor mappings of the MSIX or PBA
-         * related areas, as dom0 is capable of moving the position of the BARs
-         * in the host address space.
-         *
-         * We rely on being called with the vPCI lock held once the domain is
-         * running, so the maps are not in use.
-         */
-        for ( i = 0; i < ARRAY_SIZE(pdev->vpci->msix->table); i++ )
-            if ( pdev->vpci->msix->table[i] )
-            {
-                /* If there are any maps, the domain must be running. */
-                ASSERT(spin_is_locked(&pdev->vpci->lock));
-                iounmap(pdev->vpci->msix->table[i]);
-                pdev->vpci->msix->table[i] = NULL;
-            }
-    }
-
-    return 0;
-}
-
 static int cf_check init_msix(struct pci_dev *pdev)
 {
     struct domain *d = pdev->domain;
@@ -741,11 +632,9 @@ static int cf_check init_msix(struct pci_dev *pdev)
         vpci_msix_arch_init_entry(&msix->entries[i]);
     }
 
-    if ( list_empty(&d->arch.hvm.msix_tables) )
-        register_mmio_handler(d, &vpci_msix_table_ops);
-
     pdev->vpci->msix = msix;
-    list_add(&msix->next, &d->arch.hvm.msix_tables);
+
+    vpci_msix_arch_register(msix, d);
 
     return 0;
 }
